% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geometry.r
\name{calculate_intersection}
\alias{calculate_intersection}
\title{Calculate the intersection of a line segment and a node's bounding box.
With retraction to prevents arrowheads from piercing the shape.}
\usage{
calculate_intersection(x_start, y_start, x_end, y_end, node_type, w, h)
}
\arguments{
\item{x_start}{Source x}

\item{y_start}{Source y}

\item{x_end}{Target x}

\item{y_end}{Target y}

\item{node_type}{'OBSERVED' or 'LATENT'}

\item{w}{width}

\item{h}{height}
}
\value{
list(x, y)
}
\description{
Calculate the intersection of a line segment and a node's bounding box.
With retraction to prevents arrowheads from piercing the shape.
}
\keyword{internal}
