% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_prompts.r
\docType{data}
\name{GEMINI_SYSTEM_PROMPT}
\alias{GEMINI_SYSTEM_PROMPT}
\title{The system prompt for the Gemini API.}
\format{
An object of class \code{character} of length 1.
}
\usage{
GEMINI_SYSTEM_PROMPT
}
\description{
The system prompt for the Gemini API.
}
\keyword{internal}
