% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.r
\name{parse_and_prepare_data}
\alias{parse_and_prepare_data}
\title{Parse JSON data and prepare data frames for plotting.}
\usage{
parse_and_prepare_data(json_data)
}
\arguments{
\item{json_data}{A string containing JSON data or a list parsed from JSON.}
}
\value{
A list containing two data frames: 'nodes_df' and 'edges_df'.
}
\description{
Parse JSON data and prepare data frames for plotting.
}
\keyword{internal}
