% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preparar_json}
\alias{preparar_json}
\title{Manipula json de eventos y estructura del curso}
\usage{
preparar_json(json_data)
}
\arguments{
\item{json_data}{A string}
}
\value{
la longitud de la lista
}
\description{
Manipula json de eventos y estructura del curso
}
\examples{
preparar_json({{"a":3,"nombre":"Mathieu"} })
}
